Improved co creators field
