+++ cms/src/fields/collection.ts
@@ -1,5 +1,6 @@
 import {CmsField, CmsFieldBase, CmsFieldRelation} from 'netlify-cms-core';
 import {
+  CO_CREATORS,
   DESCRIPTION_FIELD,
   HIDDEN_FIELD,
   ID_FIELD,
@@ -28,6 +29,7 @@ export const COLLECTION_FIELDS: Array<CmsField> = [
   ID_FIELD,
   NAME_FIELD,
   DESCRIPTION_FIELD,
+  CO_CREATORS,
   LINK_FIELD,
   IMAGE_FIELD,
   TAGS_FIELD,
