+++ cms/src/fields/constants.ts
@@ -12,6 +12,32 @@ export const CLOUDINARY_IMAGE_CONFIG = {
   },
 };
 
+export const CLOUDINARY_CONTRIBUTOR_IMAGE_CONFIG = {
+  name: 'cloudinary',
+  config: {
+    default_transformations: [
+      [
+        {
+          transformation: 'profile_picture_square',
+        },
+      ],
+    ],
+  },
+};
+
+export const CLOUDINARY_COCREATOR_IMAGE_CONFIG = {
+  name: 'cloudinary',
+  config: {
+    default_transformations: [
+      [
+        {
+          transformation: 'cocreator_image',
+        },
+      ],
+    ],
+  },
+};
+
 export const CLOUDINARY_VIDEO_CONFIG = {
   name: 'cloudinary',
   config: {
