+++ content/src/exercises/5b69ff57-10c4-4288-afc7-a6f6ed212531.json
@@ -89,13 +89,13 @@
     "coCreators": [
       {
         "name": "Script and voice: Jenny Rickardsson",
-        "avatar_url": "https://res.cloudinary.com/cupcake-29k/image/upload/v1682602411/Images/Jenny_Rickardsson_kopia_sdodwf.jpg",
-        "link": "https://www.linkedin.com/in/jenny-rickardson-6770a959/"
+        "image": "https://res.cloudinary.com/cupcake-29k/image/upload/t_cocreator_image/v1682602411/Images/Jenny_Rickardsson_kopia_sdodwf.jpg",
+        "url": "https://www.linkedin.com/in/jenny-rickardson-6770a959/"
       },
       {
         "name": "Inspired by: Anke de Vrieze, Transitionmakers toolbox",
-        "link": "https://www.linkedin.com/in/ankedevrieze/?locale=nl_NL",
-        "avatar_url": "https://www.sustainableplaceshaping.net/wp-content/uploads/2017/11/anke-de-vrieze.jpg"
+        "image": "https://res.cloudinary.com/cupcake-29k/image/upload/t_cocreator_image/v1697716381/Images/anke-de-vrieze_q8h8tj.jpg",
+        "url": "https://www.linkedin.com/in/ankedevrieze/?locale=nl_NL"
       }
     ],
     "duration": 40,
